ml/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-reflection>=1.60.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
